Clean up <UploadImage/>, rename <UploadedFile/>
