--- src/components/uploaded-file/index.ts
@@ -1,1 +1,0 @@
-export * from "./uploaded-file";
